HW Lesson 5 | Реализовал запрос с сервера Yandex погоды, отобразил на экране погоду для каждого города из списка. Спрятал ключ в BuildConfig

+++ app/src/main/java/com/example/kotlinweatherapplication/repository/InfoDTO.kt
@@ -5,11 +5,11 @@ import com.google.gson.annotations.SerializedName
 data class InfoDTO(
 
     @field:SerializedName("lon")
-    val lon: Int,
+    val lon: Double,
 
     @field:SerializedName("url")
     val url: String,
 
     @field:SerializedName("lat")
-    val lat: Int
+    val lat: Double
 )
